amendments on Niko's improvements
Notably:

- make sure it is unambiguous that authors should follow JIE's APA citation
format, not any random suggested format by data provider.

- At present, I would not state that Data Descriptors are one among many
types of articles to be hosted in a new data section. My current
understanding is that this new section is there specifically for that
new type of article. Open to changing my mind though!

- I am actually not sure all data descriptors will automatically win
gold/gold. To be discussed. But in any case, we don't actually need to
make this statement in guidelines.

 On branch master
 Your branch is up to date with 'origin/master'.

 Changes to be committed:
	modified:   authorGuidelines_additions.md
	modified:   dataDescriptor_justification.md
	modified:   opennessBadge_guidelines.md